Update main.php
I solved the where the hell did the plugin page go after install problem.  The register_activation_hooks seem to work in this file or basically not inside the class set's constructor. It was also not adding the page because there was one in the trash can that wasn't permanently deleted.